refactor: integrity as a library
